+++ mobile/android/android-components/components/service/glean/src/main/java/mozilla/components/service/glean/storages/ExperimentsStorageEngine.kt
@@ -130,7 +130,6 @@ internal object ExperimentsStorageEngine : StorageEngine {
     override val sendAsTopLevelField: Boolean
         get() = true
 
-    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
     override fun clearAllStores() {
         experiments.clear()
     }