+++ mobile/android/android-components/docs/changelog.md
@@ -181,6 +181,8 @@ permalink: /changelog/
    * ⚠️ **This is a breaking API change**: Custom pings must be explicitly
      registered with Glean at startup time. See
      `components/service/glean/docs/pings/custom.md` for more information.
+   * Disabling telemetry through `setUploadEnabled` now clears all metrics
+     immediately.
 
 * **ui-autocomplete**
   * Added an optional `shouldAutoComplete` boolean to `setText` which is currently used by `updateUrl` in `EditToolbar`.